xtest: use imported OpenSSL
This commit modifies the xtest Makefiles to deal with the cases where
xtest needs to use OpenSSL:

- When the GlobalPlatform tests are enabled (CFG_GP_PACKAGE_PATH is set),
- When the MbedTLS tests are enabled (CFG_TA_MBEDTLS=y).

For the first case, it is the "make patch" target that takes care of
copying the header files under host/xtest/for_gp/include/openssl.
For the second case however, the copy does not occur and only the -I flag
is added by [1], resulting in the following error:

  cc1: error: for_gp/include: No such file or directory [-Werror=missing-include-dirs]

Address this problem by using the imported copy of OpenSSL.

The CMake and Android builds are not modified.

Fixes: [1] 43d58a5fce8f ("xtest: Makefile: link against OpenSSL if MBed TLS is enabled in TA")
Fixes: https://github.com/OP-TEE/optee_test/issues/285
Reported-by: Igor Opaniuk <igor.opaniuk@linaro.org>
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Tested-by: Igor Opaniuk <igor.opaniuk@linaro.org> (poplar)