regression: 4003: use bigger plain text in AES CTR test
The AES CTR test added in the parent commit results in the accelerated
crypto code being called with 3 blocks of data (48 bytes), which revealed
a bug in the 32-bit version of ce_aes_ctr_encrypt() [1].

Add some more data to exercise a similar loop in the 64-bit version of
ce_aes_ctr_encrypt(). The loop deals with 4 blocks of data.

Link: [1] https://github.com/OP-TEE/optee_os/issues/2305
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>